Fix voice speaking statuses
Somewhere along the line, Discord servers stopped filling the `d.ssrc` field of Voice Gateway OP 5
Speaking packets with the SSRC of the sender. However, instead of rejecting packets with missing
`d.ssrc`, they started filling it with what seems to be a SSRC for a new ghost user (increments
sequentially based on the highest existing SSRC).

Co-authored-by: zaza7 <zaza7@users.noreply.github.com>