Detect Retry-After header resolution
Discord breaks RFC here and uses milliseconds instead of seconds (╯°□°）╯︵ ┻━┻
Cloudflare follows RFC, so some ratelimiting situations may result in request spam as a result of
treating X seconds as X milliseconds.

A Discord developer has recommended detecting the presence of a Google Cloud Load Balancer header to
determine if this header is in milliseconds and not seconds.