Optional Timeout argument to the delete function
This just adds a timeout option where you specify the delay in ms for a message to be deleted. This makes it easier in general to set timeouts/delays without having to call setTimeout manually. This may not be the best way to implement it either, but it gets the job done