Acceptance of all alphabetic characters in slash commands.
In the names of slash commands, only the letters of the English alphabet were accepted. However, Discord actually supports all alphabets, but this was being blocked by Eris. For example, in Turkish, the command "/zamanaşımı" (timeout) was being prevented from being added due to the letters "ı" and "ş". Similarly, when creating a Russian command like "/таймаут" (timeout), it was being blocked in the same way. However, Discord actually allows the creation of all these commands.